Used a hacked way for registering CytoPanelComponent
An index exception related to drawing of JTabbedPane for CytoPane is
thrown often when the bundle is update or the app is closed for tabs
used to display pathway related things (e.g. diagram, PGM). Directly add
tabs into JTabbedPane. This may reduce the chance of exception thrown.
But occasionally this exception may be thrown still. This is more like a
threading issue in the implementation of Cytoscape.